+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
 
-## 0.6.1
+## 1.0.0
+
+**BREAKING CHANGES**
+- `Request.body` is used instead of `bodyStream` and `bodyBytes`.
+- `Stream<List<int>>` inputs are no longer supported, use restartable `FutureOr<Stream<List<int>>> StreamFn()` instead.
+
+Other updates
 
 - Explicit expiration in `UpdatingClient`.
 - Support form-encoded values as request body.
+- Support `File` with `Request.body` on `ConsoleClient`.
+- Support native request types with `Request.body` on `BrowserClient`.
 
 ## 0.6.0